Separate JWT logic in another file
