+++ backEnd/authServer.ts
@@ -1,26 +1,22 @@
 import dotenv from 'dotenv'
 dotenv.config()
 
-import redis from 'redis'
-import jwt from 'jsonwebtoken'
 import express from 'express'
 
 import { v4 as uuid } from 'uuid'
 import bcrypt from 'bcrypt'
 
-import mysqlConnection from './mysqlConnection'
+import mysqlConnection from './MySQLConnection'
+
+import { 
+    verifyAccessToken, verifyRefreshToken, logoutToken,
+    generateAccessToken, generateRefreshToken
+        } from './authHelpers'
 
 const 
     app = express(),
     port = 8082,
-    client = redis.createClient(6379),
-    sendQuery = mysqlConnection.getInstance().sendQuery,
-
-    redisAccessTokenExpiration  = 5,
-    redisRefreshTokenExpiration = 20,
-
-    accessTokenExpiration = redisAccessTokenExpiration.toString() + 's',
-    refreshTokenExpiration = redisRefreshTokenExpiration.toString() + 's'
+    sendQuery = mysqlConnection.getInstance().sendQuery
 
 
 app.use(express.json())
@@ -29,152 +25,7 @@ app.use(express.urlencoded({
 }))
 
 
-const checkBlacklistToken = async (token: string) => {
-    return new Promise((resolve, reject) => {
-        client.exists(token, (err, ans) => {
-            if (err)
-                reject(err)
-            else
-                resolve(!ans)
-        })
-    })
-}
 
-const generateAccessToken = (user: Object) => {
-    return jwt.sign(user, process.env.accessTokenSecret, { expiresIn: accessTokenExpiration })
-}
-
-const generateRefreshToken = (user: Object) => {
-    return jwt.sign(user, process.env.refreshTokenSecret, { expiresIn: refreshTokenExpiration })
-}
-
-const verifyAccessToken = async (req: express.Request, res: express.Response, next: Function) => {
-
-    const authTokens = req.headers['authorization'].split(' ')
-
-    if (authTokens[1] == 'null')
-        return res.sendStatus(403)
-
-    const accessToken = authTokens[1]
-
-    if (!accessToken)
-        return res.sendStatus(403)
-
-    const validAccessToken = await checkBlacklistToken(accessToken)
-
-    if (!validAccessToken)
-        return res.sendStatus(403)
-
-    jwt.verify(accessToken, process.env.accessTokenSecret, (err: jwt.VerifyErrors, user: object) => {
-        if (err) {
-            if (err instanceof jwt.TokenExpiredError)
-                return res.sendStatus(401)
-            return res.sendStatus(403)
-        }
-
-        delete user['iat']
-        delete user['exp']
-
-        if (next)
-            return next()
-
-        return res.json({
-            user: user
-        })
-    })
-}
-
-const verifyRefreshToken = async (req: express.Request, res: express.Response) => {
-
-    const authTokens = req.headers['authorization'].split(' ')
-
-    if (authTokens[1] == 'null')
-        return res.sendStatus(403)
-
-    const refreshToken = authTokens[1]
-
-    if (!refreshToken)
-        return res.sendStatus(403)
-
-    const validRefreshToken = await checkBlacklistToken(refreshToken)
-
-    if (!validRefreshToken)
-        return res.sendStatus(403)
-
-    jwt.verify(refreshToken, process.env.refreshTokenSecret, (err, user) => {
-        if (err) {
-            if (err instanceof jwt.TokenExpiredError) {
-                // const payLoad = jwt.verify(refreshToken, process.env.refreshTokenSecret, { ignoreExpiration: true })
-
-                // logout(payLoad.id)
-            }
-
-            return res.sendStatus(403)  
-        }
-
-        delete user['iat']
-        delete user['exp']
-        
-        const accessToken = generateAccessToken(user)
-
-        const userInfo = jwt.verify(accessToken, process.env.accessTokenSecret)
-
-        delete userInfo['iat']
-        delete userInfo['exp']
-
-        return res.status(200).json({
-            user: userInfo,
-            accessToken: accessToken
-        })
-    })
-}
-
-const logoutToken = async (req: express.Request, res: express.Response) => {
-
-    const authTokens = req.headers['authorization'].split(' ')
-
-    if (authTokens[1] == 'null' || authTokens[2] == 'null')
-        return res.sendStatus(403)
-
-    const accessToken = authTokens[1], refreshToken = authTokens[2]
-
-    if (!(accessToken && refreshToken))
-        return res.sendStatus(403)
-
-    const validRefreshToken = await checkBlacklistToken(refreshToken), validAccessToken = await checkBlacklistToken(accessToken)
-
-    if (!(validRefreshToken && validAccessToken))
-        return res.sendStatus(403)
-    
-    let done = false
-
-    jwt.verify(accessToken, process.env.accessTokenSecret, (err: jwt.VerifyErrors, user: object) => {
-        if (err) {
-            done = true
-            if (err instanceof jwt.TokenExpiredError)
-                return res.sendStatus(401)
-            return res.sendStatus(403)
-        }
-    })
-
-    if (done)
-        return
-
-    jwt.verify(refreshToken, process.env.refreshTokenSecret, (err, user) => {
-        if (err)
-            return res.sendStatus(403)
-
-        client.setex(accessToken, redisAccessTokenExpiration, accessToken)
-        client.setex(refreshToken, redisRefreshTokenExpiration, refreshToken)
-
-
-        // const payLoad = jwt.verify(refreshToken, process.env.refreshTokenSecret, { ignoreExpiration: true })
-
-        // logout(payLoad.id)
-    
-        return res.sendStatus(200)
-    })
-}
 
 
 app.post('/verifyAccessToken', verifyAccessToken, async (req, res) => {
@@ -182,6 +33,10 @@ app.post('/verifyAccessToken', verifyAccessToken, async (req, res) => {
 })
 
 app.post('/verifyRefreshToken', verifyRefreshToken, async (req, res) => {
+    return res.sendStatus(200)
+})
+
+app.post('/logout', logoutToken, async (req, res) => {
     return res.sendStatus(200)
 })
 
@@ -265,8 +120,7 @@ app.post('/signin', async (req, res) => {
         return res.json(error('Wrong password.'))
 
     const userData = {
-        username,
-        message: 'esti super tare'
+        username
     }
 
     const accessToken = generateAccessToken(userData)
@@ -276,8 +130,6 @@ app.post('/signin', async (req, res) => {
         accessToken,
         refreshToken
     }))
-
-    return res.json(success())
 })
 
 app.post('/checkUsernameTaken', async (req, res) => {
